+++ src/presentation/helpers/validators/required-field-validation.ts
@@ -1,4 +1,4 @@
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 import { MissingParamError } from '../../errors'
 
 export class RequiredFieldValidation implements Validation {
