+++ src/presentation/helpers/validators/validation-composite.ts
@@ -1,4 +1,4 @@
-import { Validation } from './validation'
+import { Validation } from '../../protocols/validation'
 
 export class ValidationComposite implements Validation {
   private readonly validations: Validation[]
